fix: remove dead imports and code from wave2 merges

+++ app/src/pricing/full_impl.rs
@@ -35,8 +35,8 @@ pub struct PlanPricing {
 /// Placeholder for the former GraphQL `AddonCreditsOption` type.
 #[derive(Debug, Clone)]
 pub struct AddonCreditsOption {
-    pub credits: i32,
-    pub price_usd_cents: i32,
+    pub credits: u64,
+    pub price_usd_cents: u64,
 }
 
 impl AddonCreditsOption {
@@ -87,3 +87,4 @@ impl Entity for PricingInfoModel {
 }
 
 impl SingletonEntity for PricingInfoModel {}
+
